+++ sdk/storage/azure-storage-blob/tests/test_common_blob_async.py
@@ -217,13 +217,14 @@ class TestStorageCommonBlobAsync(AsyncStorageRecordedTestCase):
             # Assert
             assert destination_blob_data == source_data
         finally:
-            async with aiohttp.ClientSession() as session:
-                await session.delete(
-                    url=base_url,
-                    headers=_build_base_file_share_headers(bearer_token_string, 0),
-                    params={'restype': 'share'}
-                )
-            await blob_service_client.delete_container(self.source_container_name)
+            if self.is_live:
+                async with aiohttp.ClientSession() as session:
+                    await session.delete(
+                        url=base_url,
+                        headers=_build_base_file_share_headers(bearer_token_string, 0),
+                        params={'restype': 'share'}
+                    )
+                await blob_service_client.delete_container(self.source_container_name)
 
     @BlobPreparer()
     @recorded_by_proxy_async
